+++ CHANGES.rst
@@ -70,3 +70,5 @@ Changelog
 - Fixed bug in get_argument_to_typevar.
 - Added new property type for configurations, MapProperty, which behaves like a dict from
   strings to some sub-property type.
+- Configurations now support: validation of additional properties, initialisation by value
+  or JSON, programmatically getting/setting additional properties.